app/platform/fabric/e2e-test/requirement.txt
#
#    SPDX-License-Identifier: Apache-2.0
#
behave
nose
flask==0.10.1
python-dateutil==2.2
pytz==2014.3
pyyaml==3.10
couchdb==1.0
flask-cors==2.0.1
requests
pyOpenSSL==16.2.0
pysha3==1.0b1
six
grpcio==0.13.1
ecdsa
python-slugify
b3j0f.aop
google
protobuf
pykafka
pyexecjs
cython
psutil
assertpy
behave_rest